requirements.txt
boto3==1.18.30
botocore==1.21.30
jmespath==0.10.0
numpy
pandas
psycopg2-binary
python-dateutil
pytz
s3transfer
six
urllib3
